+++ CHANGELOG.md
@@ -5,14 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add option to enable incomplete dates
 ### Fixed
 - Fix error when cancelling game selection
 
 ## [1.0.1] - 2019-11-25
-### Fixed
-- [Playnite #3](https://playnite.link/forum/thread-24-post-49.html#pid49) Incomplete release dates will now longer crash metadata downloading
 ### Changed
 - Tags without a matching name will now be ignored, instead of being shown as "UNKNOWN TAG"
+### Fixed
+- [Playnite #3](https://playnite.link/forum/thread-24-post-49.html#pid49) Incomplete release dates will now longer crash metadata downloading
 
 ## [1.0] - 2019-11-24
 ### Fixed